requirements.txt
Flask==0.10.1
Flask-Bootstrap==3.0.3.1
Flask-Login==0.2.7
Flask-Mail==0.9.0
Flask-Migrate==1.0.0
Flask-OAuth==0.12
Flask-Principal==0.4.0
Flask-SQLAlchemy==1.0
Flask-Script==0.6.6
Flask-Security==1.6.9
Flask-Social==1.6.2
Flask-WTF==0.9.3
Jinja2==2.7.1
Mako==0.9.0
MarkupSafe==0.18
SQLAlchemy==0.8.4
WTForms==1.0.5
Werkzeug==0.9.4
alembic==0.6.1
blinker==1.3
flask-heroku==0.1.8
gunicorn==18.0
httplib2==0.8
itsdangerous==0.23
oauth2==1.5.211
oauthlib==0.6.0
passlib==1.6.1
psycopg2==2.5.1
python-dateutil==2.2
python-twitter==1.1
requests==2.1.0
requests-oauthlib==0.4.0
simplejson==3.3.1
six==1.4.1
wsgiref==0.1.2
